Don’t set fetch-depth to 0
According to the documentation for actions/checkout, this is what
fetch-depth does [1]:

>    # Number of commits to fetch. 0 indicates all history for all branches and tags.
>    # Default: 1
>    fetch-depth: ''

Before this commit, we set fetch-depth to 0. As far as I can tell, we
don’t need any of the repo’s history in order to deploy the Web site. I
ran “yarn build” locally and inspected the build/ directory. I wasn’t
able to find any references to the hash of the commit that I used to
build the site.

Additionally, I wrote this script:

    #!/usr/bin/env bash
    set -e

    program_name="$(basename --suffix=.sh "$0")"
    readonly program_name
    dest="$(mktemp --directory --suffix="$program_name")"
    readonly dest

    echo -E "This script will make a copy of the repo at “$dest”."
    echo "Don’t forget to delete that folder after you’re done."
    echo

    git archive --format=tar HEAD | tar --extract --directory="$dest"
    cd "$dest"
    yarn install --frozen-lockfile
    yarn build

    echo
    echo -E "Done building the site. The contents are at “$dest/build”."

If you cd into a local copy of the smartkidsllc.github.io repo and then
run that script, then it will copy the site’s source code into a
temporary directory and build the site. Importantly, that script does
not copy any Git metadata into the temporary directory. In other words,
I was able to use that script in order to build the site without any of
the repo’s commit history present. The site builds and works fine for me
when I run that script.

Based on the testing that I did locally, I don’t really see a benefit to
downloading all of the repository’s history when building the site in
GitHub Actions. It just seems like a waste of time.

[1]: <https://github.com/actions/checkout/tree/v4.1.7?tab=readme-ov-file#usage>