Don’t try to deploy forks of this repo
Recently, I pushed to the main branch of my fork of
smartkidsllc.github.io. I then got an email from GitHub telling me that
the Deploy to GitHub Pages job had failed [1]. This is to be expected
because I don’t have GitHub Pages enabled on my fork.

I don’t actually want my fork to be deployed. The fork only exists so
that I can open pull requests without creating branches on the main
repository. I think that all (if not most) of the forks of this repo
that will ever be created will end up being similar. They’ll all exist
just so that people can create pull requests.

This change makes it so that the Deploy to GitHub Pages job will only be
run on the original repository. Doing so will make it so that people
don’t get errors when they push to their fork’s main branch.

Credit goes to @s-weigand for figuring out how to prevent jobs from
being run in forks [2]. His solution uses github.repository_owner. I
decided to use github.repository_owner_id instead.
github.repository_owner is the username of a GitHub account, and
github.repository_owner_id is the account ID of a GitHub account [3]. I
thought that it would be better to use the account ID because the
account ID is less likely to change than the account name [4][5].

[1]: <https://github.com/Jayman2000/smartkidsllc.github.io-pr/actions/runs/10563804100/job/29264950067>
[2]: <https://github.com/orgs/community/discussions/26409#discussioncomment-3251822>
[3]: <https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/accessing-contextual-information-about-workflow-runs#github-context>
[4]: <https://stackoverflow.com/q/68049001/7593853>
[5]: <https://docs.github.com/en/rest/users/users?apiVersion=2022-11-28#get-a-user-using-their-id>