Make sure that all files end with a newline
Much has already been said about why text files should end with a
newline [1]. I’m sure that a lot has been said about why text files
should not end with a newline as well. I don’t know if it’s better for
text files to end with a newline or to not end with a newline. I just
know that before this commit the majority of the text files in this repo
ended with a newline. This commit makes all of the text files in this
repo consistent by making sure that they all end with a newline.

The main motivation behind this change is to make it easier to create a
future commit. That future commit will add an EditorConfig file to this
repo. That EditorConfig file will have a insert_final_newline property.
It will be easier to specify the insert_final_newline property correctly
now that all files end with a newline. Otherwise, I would have to set
insert_final_newline property to true for some files and to false for
other files.

[1]: <https://stackoverflow.com/q/729692/7593853>