Create Nix flake for working on this repo
Unlike most package managers, the Nix package manager allows you to pin
revisions of package repositories. For example, let’s say that I’m using
nixpkgs (Nix’s official package repository). I could pin the nixpkgs
revision to 36bae45077667aff5720e5b3f1a5458f51cf0776. Then, I could
install the ecwolf package. When I do that, the Nix package manager
guarantees the following:

1. I’m guaranteed to get version 1.4.1 of the ecwolf package.
2. ecwolf depends on SDL2. I’m guaranteed to get a version of ecwolf
   that was built with version 2.30.2 of SDL2.
3. SDL2 depends on libX11. I’m guaranteed to get a version of SDL2 that
   was built with version 1.8.9 of libX11.
4. libX11 depends on…

The point is: the Nix package manage offers strong guarantees about the
versions of packages that get used. As far as I know, the only package
managers that offer guarantees that are that strong are the Nix package
manager and forks of the Nix package manager.

---

This commit adds a flake.nix file that guarantees that a known-good
version of Node.js is used. This will allow anyone who uses the Nix
package manager to set up a development environment for this repo by
running a single command.

Additionally, you need to use a special workaround in order to get
corepack to work properly on NixOS. In a future commit, I’m going to
rewrite the instructions for testing out this repo locally, and those
new instructions are going to depend on corepack. Adding a flake.nix
file now will make it easier to write those instructions in the future.
Specifically, it will make it easier to write the part of the
instructions that tell NixOS users how to enable corepack.