Remove periods from the ends of descriptions
Before this change, some of the category descriptions and some of the
page descriptions ended with a period, and some of them didn’t. It
doesn’t really make sense for category and page descriptions to end with
a period because they aren’t complete sentences.

This commit makes the repository more consistent by making sure that
none of the descriptions end with a period.