Rename update-script to update-script-part-1
The idea here is to eventually have two update scripts.
update-script-part-1 will do all of the updates that are related to Nix,
and update-script-part-2 will do all of the updates that aren’t related
to Nix. There’s two reasons why I want to do this:

1. After flake.nix and flake.lock are updated, we should really invoke
nix a second time in order to do the rest of the updates. That way the
rest of the updates are done using the new version of the flake rather
than the old version of the flake.

2. I want to make it so that update-script-part-2 can be run even if you
don’t have the Nix package manager. You don’t need to have the Nix
package manager in order to use this repo, so you shouldn’t need to have
the Nix package manager in order to update this repo. At least, you
shouldn’t need to have the Nix package manager in order to update the
parts of this repo that have nothing to do with Nix.