Make update-script-part-2 update package versions in package.json
In package.json, we use SemVer ranges [1] when specifying the npm
packages that this project depends on. For example, we use the SemVer
range "^18.0.0" when specifying that we depend on the react package.

Before this change, there was no automated process for updating those
SemVer ranges or even for notifying us that those SemVer ranges need to
be updated. This meant that we would have to remember to do certain
updates manually. For example, version 19.2.1 of the react package is
available [2]. That version falls outside of the previously mentioned
"^18.0.0" SemVer range. Before this change, we would have had to
remember to manually update that SemVer range or else we would never
start using version 19.2.1 of the react package.

This change fixes that issue by making update-script-part-2
automatically update the SemVer ranges in package.json.

Closes #63.

[1]: <https://docs.npmjs.com/cli/v6/using-npm/semver#ranges>
[2]: <https://www.npmjs.com/package/react/v/19.2.1>